Split tanjun.commands into directory + fix previous tests
