+++ tanjun/commands/__init__.py
@@ -29,11 +29,45 @@
 # CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-"""Standard implementations of Tanjun's context classes."""
+"""Standard implementations of Tanjun's command objects."""
 from __future__ import annotations
 
-__all__: list[str] = ["MessageContext", "SlashContext", "SlashOption"]
+__all__: list[str] = [
+    "as_message_command",
+    "as_message_command_group",
+    "as_slash_command",
+    "slash_command_group",
+    "MessageCommand",
+    "MessageCommandGroup",
+    "BaseSlashCommand",
+    "SlashCommand",
+    "SlashCommandGroup",
+    "with_str_slash_option",
+    "with_int_slash_option",
+    "with_float_slash_option",
+    "with_bool_slash_option",
+    "with_role_slash_option",
+    "with_user_slash_option",
+    "with_member_slash_option",
+    "with_channel_slash_option",
+    "with_mentionable_slash_option",
+]
 
-from .message import MessageContext
-from .slash import SlashContext
-from .slash import SlashOption
+from .message import MessageCommand
+from .message import MessageCommandGroup
+from .message import as_message_command
+from .message import as_message_command_group
+from .slash import BaseSlashCommand
+from .slash import SlashCommand
+from .slash import SlashCommandGroup
+from .slash import as_slash_command
+from .slash import slash_command_group
+from .slash import with_bool_slash_option
+from .slash import with_channel_slash_option
+from .slash import with_float_slash_option
+from .slash import with_int_slash_option
+from .slash import with_member_slash_option
+from .slash import with_mentionable_slash_option
+from .slash import with_role_slash_option
+from .slash import with_str_slash_option
+from .slash import with_user_slash_option
