+++ tanjun/commands/menu.py
@@ -29,11 +29,7 @@
 # CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-"""Standard implementations of Tanjun's context classes."""
+"""Standard implementation of Tanjun's command objects."""
 from __future__ import annotations
 
-__all__: list[str] = ["MessageContext", "SlashContext", "SlashOption"]
-
-from .message import MessageContext
-from .slash import SlashContext
-from .slash import SlashOption
+__all__: list[str] = []
