+++ tanjun/commands/slash.py
@@ -33,15 +33,9 @@
 from __future__ import annotations
 
 __all__: list[str] = [
-    "AnyMessageCommandT",
     "ConverterSig",
-    "as_message_command",
-    "as_message_command_group",
     "as_slash_command",
     "slash_command_group",
-    "MessageCommand",
-    "MessageCommandGroup",
-    "PartialCommand",
     "BaseSlashCommand",
     "SlashCommand",
     "SlashCommandGroup",
@@ -54,6 +48,7 @@ __all__: list[str] = [
     "with_member_slash_option",
     "with_channel_slash_option",
     "with_mentionable_slash_option",
+    "UNDEFINED_DEFAULT",
 ]
 
 import copy
@@ -64,21 +59,18 @@ from collections import abc as collections
 
 import hikari
 
-from . import abc
-from . import checks as checks_
-from . import components
-from . import conversion
-from . import errors
-from . import hooks as hooks_
-from . import injecting
-from . import utilities
+from .. import abc
+from .. import components
+from .. import conversion
+from .. import errors
+from .. import hooks as hooks_
+from .. import injecting
+from .. import utilities
+from . import base
 
 if typing.TYPE_CHECKING:
     from hikari.api import special_endpoints as special_endpoints_api
 
-    _MessageCommandT = typing.TypeVar("_MessageCommandT", bound="MessageCommand[typing.Any]")
-    _MessageCommandGroupT = typing.TypeVar("_MessageCommandGroupT", bound="MessageCommandGroup[typing.Any]")
-    _PartialCommandT = typing.TypeVar("_PartialCommandT", bound="PartialCommand[typing.Any]")
     _BaseSlashCommandT = typing.TypeVar("_BaseSlashCommandT", bound="BaseSlashCommand")
     _SlashCommandT = typing.TypeVar("_SlashCommandT", bound="SlashCommand[typing.Any]")
     _SlashCommandGroupT = typing.TypeVar("_SlashCommandGroupT", bound="SlashCommandGroup")
@@ -90,93 +82,10 @@ _CallbackishT = typing.Union[
     abc.SlashCommand[abc.CommandCallbackSigT],
 ]
 
-AnyMessageCommandT = typing.TypeVar("AnyMessageCommandT", bound=abc.MessageCommand[typing.Any])
 ConverterSig = collections.Callable[..., abc.MaybeAwaitableT[typing.Any]]
 """Type hint of a converter used for a slash command option."""
 _EMPTY_DICT: typing.Final[dict[typing.Any, typing.Any]] = {}
 _EMPTY_HOOKS: typing.Final[hooks_.Hooks[typing.Any]] = hooks_.Hooks()
-
-
-class PartialCommand(abc.ExecutableCommand[abc.ContextT], components.AbstractComponentLoader):
-    """Base class for the standard ExecutableCommand implementations."""
-
-    __slots__ = ("_checks", "_component", "_hooks", "_metadata")
-
-    def __init__(self) -> None:
-        self._checks: list[checks_.InjectableCheck] = []
-        self._component: typing.Optional[abc.Component] = None
-        self._hooks: typing.Optional[abc.Hooks[abc.ContextT]] = None
-        self._metadata: dict[typing.Any, typing.Any] = {}
-
-    @property
-    def checks(self) -> collections.Collection[abc.CheckSig]:
-        # <<inherited docstring from tanjun.abc.ExecutableCommand>>.
-        return tuple(check.callback for check in self._checks)
-
-    @property
-    def component(self) -> typing.Optional[abc.Component]:
-        # <<inherited docstring from tanjun.abc.ExecutableCommand>>.
-        return self._component
-
-    @property
-    def hooks(self) -> typing.Optional[abc.Hooks[abc.ContextT]]:
-        # <<inherited docstring from tanjun.abc.ExecutableCommand>>.
-        return self._hooks
-
-    @property
-    def metadata(self) -> collections.MutableMapping[typing.Any, typing.Any]:
-        # <<inherited docstring from tanjun.abc.ExecutableCommand>>.
-        return self._metadata
-
-    @property
-    def needs_injector(self) -> bool:
-        # <<inherited docstring from tanjun.injecting.Injectable>>.
-        return any(check.needs_injector for check in self._checks)
-
-    def copy(self: _PartialCommandT, *, _new: bool = True) -> _PartialCommandT:
-        # <<inherited docstring from tanjun.abc.ExecutableCommand>>.
-        if not _new:
-            self._checks = [check.copy() for check in self._checks]
-            self._hooks = self._hooks.copy() if self._hooks else None
-            self._metadata = self._metadata.copy()
-            return self
-
-        return copy.copy(self).copy(_new=False)
-
-    def set_hooks(self: _PartialCommandT, hooks: typing.Optional[abc.Hooks[abc.ContextT]], /) -> _PartialCommandT:
-        # <<inherited docstring from tanjun.abc.ExecutableCommand>>.
-        self._hooks = hooks
-        return self
-
-    def set_metadata(self: _PartialCommandT, key: typing.Any, value: typing.Any, /) -> _PartialCommandT:
-        # <<inherited docstring from tanjun.abc.ExecutableCommand>>.
-        self._metadata[key] = value
-        return self
-
-    def add_check(self: _PartialCommandT, check: abc.CheckSig, /) -> _PartialCommandT:
-        # <<inherited docstring from tanjun.abc.ExecutableCommand>>.
-        if check not in self._checks:
-            self._checks.append(checks_.InjectableCheck(check))
-
-        return self
-
-    def remove_check(self: _PartialCommandT, check: abc.CheckSig, /) -> _PartialCommandT:
-        # <<inherited docstring from tanjun.abc.ExecutableCommand>>.
-        self._checks.remove(typing.cast("checks_.InjectableCheck", check))
-        return self
-
-    def with_check(self, check: abc.CheckSigT, /) -> abc.CheckSigT:
-        self.add_check(check)
-        return check
-
-    def bind_client(self: _PartialCommandT, client: abc.Client, /) -> _PartialCommandT:
-        # <<inherited docstring from tanjun.abc.ExecutableCommand>>.
-        return self
-
-    def bind_component(self: _PartialCommandT, component: abc.Component, /) -> _PartialCommandT:
-        # <<inherited docstring from tanjun.abc.ExecutableCommand>>.
-        self._component = component
-        return self
 
 
 _SCOMMAND_NAME_REG: typing.Final[re.Pattern[str]] = re.compile(r"^[\w-]{1,32}$", flags=re.UNICODE)
@@ -396,7 +305,8 @@ def as_slash_command(
     return decorator
 
 
-_UNDEFINED_DEFAULT = object()
+UNDEFINED_DEFAULT = object()
+"""Singleton used for marking slash command defaults as undefined."""
 
 
 def with_str_slash_option(
@@ -406,7 +316,7 @@ def with_str_slash_option(
     *,
     choices: typing.Union[collections.Mapping[str, str], collections.Sequence[str], None] = None,
     converters: typing.Union[collections.Sequence[ConverterSig], ConverterSig] = (),
-    default: typing.Any = _UNDEFINED_DEFAULT,
+    default: typing.Any = UNDEFINED_DEFAULT,
     pass_as_kwarg: bool = True,
 ) -> collections.Callable[[_SlashCommandT], _SlashCommandT]:
     """Add a string option to a slash command.
@@ -445,7 +355,7 @@ def with_int_slash_option(
     *,
     choices: typing.Optional[collections.Mapping[str, int]] = None,
     converters: typing.Union[collections.Collection[ConverterSig], ConverterSig] = (),
-    default: typing.Any = _UNDEFINED_DEFAULT,
+    default: typing.Any = UNDEFINED_DEFAULT,
     min_value: typing.Optional[int] = None,
     max_value: typing.Optional[int] = None,
     pass_as_kwarg: bool = True,
@@ -489,7 +399,7 @@ def with_float_slash_option(
     always_float: bool = True,
     choices: typing.Optional[collections.Mapping[str, float]] = None,
     converters: typing.Union[collections.Collection[ConverterSig], ConverterSig] = (),
-    default: typing.Any = _UNDEFINED_DEFAULT,
+    default: typing.Any = UNDEFINED_DEFAULT,
     min_value: typing.Optional[float] = None,
     max_value: typing.Optional[float] = None,
     pass_as_kwarg: bool = True,
@@ -527,7 +437,7 @@ def with_float_slash_option(
 
 
 def with_bool_slash_option(
-    name: str, description: str, /, *, default: typing.Any = _UNDEFINED_DEFAULT, pass_as_kwarg: bool = True
+    name: str, description: str, /, *, default: typing.Any = UNDEFINED_DEFAULT, pass_as_kwarg: bool = True
 ) -> collections.Callable[[_SlashCommandT], _SlashCommandT]:
     """Add a boolean option to a slash command.
 
@@ -551,7 +461,7 @@ def with_bool_slash_option(
 
 
 def with_user_slash_option(
-    name: str, description: str, /, *, default: typing.Any = _UNDEFINED_DEFAULT, pass_as_kwarg: bool = True
+    name: str, description: str, /, *, default: typing.Any = UNDEFINED_DEFAULT, pass_as_kwarg: bool = True
 ) -> collections.Callable[[_SlashCommandT], _SlashCommandT]:
     """Add a user option to a slash command.
 
@@ -580,7 +490,7 @@ def with_user_slash_option(
 
 
 def with_member_slash_option(
-    name: str, description: str, /, *, default: typing.Any = _UNDEFINED_DEFAULT
+    name: str, description: str, /, *, default: typing.Any = UNDEFINED_DEFAULT
 ) -> collections.Callable[[_SlashCommandT], _SlashCommandT]:
     """Add a member option to a slash command.
 
@@ -633,7 +543,7 @@ def with_channel_slash_option(
     /,
     *,
     types: typing.Union[collections.Collection[type[hikari.PartialChannel]], None] = None,
-    default: typing.Any = _UNDEFINED_DEFAULT,
+    default: typing.Any = UNDEFINED_DEFAULT,
     pass_as_kwarg: bool = True,
 ) -> collections.Callable[[_SlashCommandT], _SlashCommandT]:
     """Add a channel option to a slash command.
@@ -661,7 +571,7 @@ def with_channel_slash_option(
 
 
 def with_role_slash_option(
-    name: str, description: str, /, *, default: typing.Any = _UNDEFINED_DEFAULT, pass_as_kwarg: bool = True
+    name: str, description: str, /, *, default: typing.Any = UNDEFINED_DEFAULT, pass_as_kwarg: bool = True
 ) -> collections.Callable[[_SlashCommandT], _SlashCommandT]:
     """Add a role option to a slash command.
 
@@ -685,7 +595,7 @@ def with_role_slash_option(
 
 
 def with_mentionable_slash_option(
-    name: str, description: str, /, *, default: typing.Any = _UNDEFINED_DEFAULT, pass_as_kwarg: bool = True
+    name: str, description: str, /, *, default: typing.Any = UNDEFINED_DEFAULT, pass_as_kwarg: bool = True
 ) -> collections.Callable[[_SlashCommandT], _SlashCommandT]:
     """Add a mentionable option to a slash command.
 
@@ -730,7 +640,7 @@ class _TrackedOption:
         always_float: bool = False,
         converters: typing.Optional[list[injecting.CallbackDescriptor[typing.Any]]] = None,
         only_member: bool = False,
-        default: typing.Any = _UNDEFINED_DEFAULT,
+        default: typing.Any = UNDEFINED_DEFAULT,
     ) -> None:
         self.converters = converters or []
         self.default = default
@@ -815,7 +725,7 @@ class _SlashCommandBuilder(hikari.impl.SlashCommandBuilder):
         return builder
 
 
-class BaseSlashCommand(PartialCommand[abc.SlashContext], abc.BaseSlashCommand):
+class BaseSlashCommand(base.PartialCommand[abc.SlashContext], abc.BaseSlashCommand):
     """Base class used for the standard slash command implementations."""
 
     __slots__ = ("_defaults_to_ephemeral", "_description", "_is_global", "_name", "_parent", "_tracked_command")
@@ -1280,7 +1190,7 @@ class SlashCommand(BaseSlashCommand, abc.SlashCommand[abc.CommandCallbackSigT]):
             collections.Mapping[str, typing.Union[str, int, float]], collections.Sequence[typing.Any], None
         ] = None,
         converters: typing.Union[collections.Iterable[ConverterSig], ConverterSig] = (),
-        default: typing.Any = _UNDEFINED_DEFAULT,
+        default: typing.Any = UNDEFINED_DEFAULT,
         min_value: typing.Union[int, float, None] = None,
         max_value: typing.Union[int, float, None] = None,
         only_member: bool = False,
@@ -1327,7 +1237,7 @@ class SlashCommand(BaseSlashCommand, abc.SlashCommand[abc.CommandCallbackSigT]):
         if actual_choices and len(actual_choices) > 25:
             raise ValueError("Slash command options cannot have more than 25 choices")
 
-        required = default is _UNDEFINED_DEFAULT
+        required = default is UNDEFINED_DEFAULT
         self._builder.add_option(
             hikari.CommandOption(
                 type=type_,
@@ -1359,7 +1269,7 @@ class SlashCommand(BaseSlashCommand, abc.SlashCommand[abc.CommandCallbackSigT]):
         *,
         choices: typing.Union[collections.Mapping[str, str], collections.Sequence[str], None] = None,
         converters: typing.Union[collections.Sequence[ConverterSig], ConverterSig] = (),
-        default: typing.Any = _UNDEFINED_DEFAULT,
+        default: typing.Any = UNDEFINED_DEFAULT,
         pass_as_kwarg: bool = True,
         _stack_level: int = 0,
     ) -> _SlashCommandT:
@@ -1466,7 +1376,7 @@ class SlashCommand(BaseSlashCommand, abc.SlashCommand[abc.CommandCallbackSigT]):
         *,
         choices: typing.Optional[collections.Mapping[str, int]] = None,
         converters: typing.Union[collections.Collection[ConverterSig], ConverterSig] = (),
-        default: typing.Any = _UNDEFINED_DEFAULT,
+        default: typing.Any = UNDEFINED_DEFAULT,
         min_value: typing.Optional[int] = None,
         max_value: typing.Optional[int] = None,
         pass_as_kwarg: bool = True,
@@ -1557,7 +1467,7 @@ class SlashCommand(BaseSlashCommand, abc.SlashCommand[abc.CommandCallbackSigT]):
         always_float: bool = True,
         choices: typing.Optional[collections.Mapping[str, float]] = None,
         converters: typing.Union[collections.Collection[ConverterSig], ConverterSig] = (),
-        default: typing.Any = _UNDEFINED_DEFAULT,
+        default: typing.Any = UNDEFINED_DEFAULT,
         min_value: typing.Optional[float] = None,
         max_value: typing.Optional[float] = None,
         pass_as_kwarg: bool = True,
@@ -1653,7 +1563,7 @@ class SlashCommand(BaseSlashCommand, abc.SlashCommand[abc.CommandCallbackSigT]):
         description: str,
         /,
         *,
-        default: typing.Any = _UNDEFINED_DEFAULT,
+        default: typing.Any = UNDEFINED_DEFAULT,
         pass_as_kwarg: bool = True,
     ) -> _SlashCommandT:
         r"""Add a boolean option to a slash command.
@@ -1705,7 +1615,7 @@ class SlashCommand(BaseSlashCommand, abc.SlashCommand[abc.CommandCallbackSigT]):
         description: str,
         /,
         *,
-        default: typing.Any = _UNDEFINED_DEFAULT,
+        default: typing.Any = UNDEFINED_DEFAULT,
         pass_as_kwarg: bool = True,
     ) -> _SlashCommandT:
         r"""Add a user option to a slash command.
@@ -1761,7 +1671,7 @@ class SlashCommand(BaseSlashCommand, abc.SlashCommand[abc.CommandCallbackSigT]):
         description: str,
         /,
         *,
-        default: typing.Any = _UNDEFINED_DEFAULT,
+        default: typing.Any = UNDEFINED_DEFAULT,
     ) -> _SlashCommandT:
         r"""Add a member option to a slash command.
 
@@ -1812,7 +1722,7 @@ class SlashCommand(BaseSlashCommand, abc.SlashCommand[abc.CommandCallbackSigT]):
         description: str,
         /,
         *,
-        default: typing.Any = _UNDEFINED_DEFAULT,
+        default: typing.Any = UNDEFINED_DEFAULT,
         types: typing.Optional[collections.Collection[type[hikari.PartialChannel]]] = None,
         pass_as_kwarg: bool = True,
     ) -> _SlashCommandT:
@@ -1890,7 +1800,7 @@ class SlashCommand(BaseSlashCommand, abc.SlashCommand[abc.CommandCallbackSigT]):
         description: str,
         /,
         *,
-        default: typing.Any = _UNDEFINED_DEFAULT,
+        default: typing.Any = UNDEFINED_DEFAULT,
         pass_as_kwarg: bool = True,
     ) -> _SlashCommandT:
         r"""Add a role option to a slash command.
@@ -1940,7 +1850,7 @@ class SlashCommand(BaseSlashCommand, abc.SlashCommand[abc.CommandCallbackSigT]):
         description: str,
         /,
         *,
-        default: typing.Any = _UNDEFINED_DEFAULT,
+        default: typing.Any = UNDEFINED_DEFAULT,
         pass_as_kwarg: bool = True,
     ) -> _SlashCommandT:
         r"""Add a mentionable option to a slash command.
@@ -1994,7 +1904,7 @@ class SlashCommand(BaseSlashCommand, abc.SlashCommand[abc.CommandCallbackSigT]):
         keyword_args: dict[str, typing.Union[int, float, str, hikari.User, hikari.Role, hikari.InteractionChannel]] = {}
         for tracked_option in self._tracked_options.values():
             if not (option := ctx.options.get(tracked_option.name)):
-                if tracked_option.default is _UNDEFINED_DEFAULT:
+                if tracked_option.default is UNDEFINED_DEFAULT:
                     raise RuntimeError(  # TODO: ConversionError?
                         f"Required option {tracked_option.name} is missing data, are you sure your commands"
                         " are up to date?"
@@ -2087,429 +1997,3 @@ class SlashCommand(BaseSlashCommand, abc.SlashCommand[abc.CommandCallbackSigT]):
             return super().copy(_new=_new, parent=parent)
 
         return super().copy(_new=_new, parent=parent)
-
-
-def as_message_command(
-    name: str, /, *names: str
-) -> collections.Callable[[_CallbackishT[abc.CommandCallbackSigT],], MessageCommand[abc.CommandCallbackSigT]]:
-    """Build a message command from a decorated callback.
-
-    Parameters
-    ----------
-    name : str
-        The command name.
-
-    Other Parameters
-    ----------------
-    *names : str
-        Variable positional arguments of other names for the command.
-
-    Returns
-    -------
-    collections.abc.Callable[[_CallbackishT[CommandCallbackSigT]], MessageCommand[CommandCallbackSigT]]
-        The decorator callback used to make a `MessageCommand`.
-
-        This can either wrap a raw command callback or another callable command instance
-        (e.g. `SlashCommand`, `MessageCommand`, `MessageCommandGroup`) and will manage
-        loading the other command into a component when using `tanjun.Component.load_from_scope`.
-    """
-
-    def decorator(
-        callback: _CallbackishT[abc.CommandCallbackSigT],
-        /,
-    ) -> MessageCommand[abc.CommandCallbackSigT]:
-        if isinstance(callback, (abc.SlashCommand, abc.MessageCommand)):
-            return MessageCommand(callback.callback, name, *names, _wrapped_command=callback)
-
-        return MessageCommand(callback, name, *names)
-
-    return decorator
-
-
-def as_message_command_group(
-    name: str, /, *names: str, strict: bool = False
-) -> collections.Callable[[_CallbackishT[abc.CommandCallbackSigT]], MessageCommandGroup[abc.CommandCallbackSigT]]:
-    """Build a message command group from a decorated callback.
-
-    Parameters
-    ----------
-    name : str
-        The command name.
-
-    Other Parameters
-    ----------------
-    *names : str
-        Variable positional arguments of other names for the command.
-    strict : bool
-        Whether this command group should only allow commands without spaces in their names.
-
-        This allows for a more optimised command search pattern to be used and
-        enforces that command names are unique to a single command within the group.
-
-    Returns
-    -------
-    collections.abc.Callable[[_CallbackishT[CommandCallbackSigT]], MessageCommand[CommandCallbackSigT]]
-        The decorator callback used to make a `MessageCommandGroup`.
-
-        This can either wrap a raw command callback or another callable command instance
-        (e.g. `SlashCommand`, `MessageCommand`, `MessageCommandGroup`) and will manage
-        loading the other command into a component when using `tanjun.Component.load_from_scope`.
-    """
-
-    def decorator(callback: _CallbackishT[abc.CommandCallbackSigT], /) -> MessageCommandGroup[abc.CommandCallbackSigT]:
-        if isinstance(callback, (abc.SlashCommand, abc.MessageCommand)):
-            return MessageCommandGroup(callback.callback, name, *names, strict=strict, _wrapped_command=callback)
-
-        return MessageCommandGroup(callback, name, *names, strict=strict)
-
-    return decorator
-
-
-class MessageCommand(PartialCommand[abc.MessageContext], abc.MessageCommand[abc.CommandCallbackSigT]):
-    """Standard implementation of a message command."""
-
-    __slots__ = ("_callback", "_names", "_parent", "_parser", "_wrapped_command")
-
-    def __init__(
-        self,
-        callback: abc.CommandCallbackSigT,
-        name: str,
-        /,
-        *names: str,
-        _wrapped_command: typing.Optional[abc.ExecutableCommand[typing.Any]] = None,
-    ) -> None:
-        """Initialise a message command.
-
-        Parameters
-        ----------
-        callback : collections.abc.Callable[[tanjun.abc.MessageContext, ...], collections.abc.Awaitable[None]]
-            Callback to execute when the command is invoked.
-
-            This should be an asynchronous callback which takes one positional
-            argument of type `tanjun.abc.MessageContext`, returns `None` and may use
-            dependency injection to access other services.
-        name : str
-            The command name.
-
-        Other Parameters
-        ----------------
-        *names : str
-            Variable positional arguments of other names for the command.
-        """
-        super().__init__()
-        if not _wrapped_command and isinstance(callback, (abc.MessageCommand, abc.SlashCommand)):
-            callback = typing.cast(abc.CommandCallbackSigT, callback.callback)
-
-        self._callback = injecting.CallbackDescriptor[None](callback)
-        self._names = list(dict.fromkeys((name, *names)))
-        self._parent: typing.Optional[abc.MessageCommandGroup[typing.Any]] = None
-        self._parser: typing.Optional[abc.MessageParser] = None
-        self._wrapped_command = _wrapped_command
-
-    def __repr__(self) -> str:
-        return f"Command <{self._names}>"
-
-    if typing.TYPE_CHECKING:
-        __call__: abc.CommandCallbackSigT
-
-    else:
-
-        async def __call__(self, *args, **kwargs) -> None:
-            await self._callback.callback(*args, **kwargs)
-
-    @property
-    def callback(self) -> abc.CommandCallbackSigT:
-        # <<inherited docstring from tanjun.abc.MessageCommand>>.
-        return typing.cast(abc.CommandCallbackSigT, self._callback.callback)
-
-    @property
-    # <<inherited docstring from tanjun.abc.MessageCommand>>.
-    def names(self) -> collections.Collection[str]:
-        return self._names.copy()
-
-    @property
-    def needs_injector(self) -> bool:
-        return self._callback.needs_injector
-
-    @property
-    def parent(self) -> typing.Optional[abc.MessageCommandGroup[typing.Any]]:
-        # <<inherited docstring from tanjun.abc.MessageCommand>>.
-        return self._parent
-
-    @property
-    def parser(self) -> typing.Optional[abc.MessageParser]:
-        # <<inherited docstring from tanjun.abc.MessageCommand>>.
-        return self._parser
-
-    def bind_client(self: _MessageCommandT, client: abc.Client, /) -> _MessageCommandT:
-        # <<inherited docstring from tanjun.abc.ExecutableCommand>>.
-        super().bind_client(client)
-        if self._parser:
-            self._parser.bind_client(client)
-
-        return self
-
-    def bind_component(self: _MessageCommandT, component: abc.Component, /) -> _MessageCommandT:
-        # <<inherited docstring from tanjun.abc.ExecutableCommand>>.
-        super().bind_component(component)
-        if self._parser:
-            self._parser.bind_component(component)
-
-        return self
-
-    def copy(
-        self: _MessageCommandT,
-        *,
-        parent: typing.Optional[abc.MessageCommandGroup[typing.Any]] = None,
-        _new: bool = True,
-    ) -> _MessageCommandT:
-        # <<inherited docstring from tanjun.abc.MessageCommand>>.
-        if not _new:
-            self._callback = copy.copy(self._callback)
-            self._names = self._names.copy()
-            self._parent = parent
-            self._parser = self._parser.copy() if self._parser else None
-            return super().copy(_new=_new)
-
-        return super().copy(_new=_new)
-
-    def set_parent(
-        self: _MessageCommandT, parent: typing.Optional[abc.MessageCommandGroup[typing.Any]], /
-    ) -> _MessageCommandT:
-        # <<inherited docstring from tanjun.abc.MessageCommand>>.
-        self._parent = parent
-        return self
-
-    def set_parser(self: _MessageCommandT, parser: typing.Optional[abc.MessageParser], /) -> _MessageCommandT:
-        # <<inherited docstring from tanjun.abc.MessageCommand>>.
-        self._parser = parser
-        return self
-
-    async def check_context(self, ctx: abc.MessageContext, /) -> bool:
-        # <<inherited docstring from tanjun.abc.MessageCommand>>.
-        ctx.set_command(self)
-        result = await utilities.gather_checks(ctx, self._checks)
-        ctx.set_command(None)
-        return result
-
-    async def execute(
-        self,
-        ctx: abc.MessageContext,
-        /,
-        *,
-        hooks: typing.Optional[collections.MutableSet[abc.MessageHooks]] = None,
-    ) -> None:
-        # <<inherited docstring from tanjun.abc.MessageCommand>>.
-        ctx = ctx.set_command(self)
-        own_hooks = self._hooks or _EMPTY_HOOKS
-        try:
-            await own_hooks.trigger_pre_execution(ctx, hooks=hooks)
-
-            if self._parser is not None:
-                kwargs = await self._parser.parse(ctx)
-
-            else:
-                kwargs = _EMPTY_DICT
-
-            await self._callback.resolve_with_command_context(ctx, ctx, **kwargs)
-
-        except errors.CommandError as exc:
-            response = exc.message if len(exc.message) <= 2000 else exc.message[:1997] + "..."
-            await ctx.respond(content=response)
-
-        except errors.HaltExecution:
-            raise
-
-        except Exception as exc:
-            if await own_hooks.trigger_error(ctx, exc, hooks=hooks) <= 0:
-                raise
-
-        else:
-            # TODO: how should this be handled around CommandError?
-            await own_hooks.trigger_success(ctx, hooks=hooks)
-
-        finally:
-            await own_hooks.trigger_post_execution(ctx, hooks=hooks)
-
-    def load_into_component(self, component: abc.Component, /) -> None:
-        # <<inherited docstring from tanjun.components.load_into_component>>.
-        if not self._parent:
-            component.add_message_command(self)
-
-        if self._wrapped_command and isinstance(self._wrapped_command, components.AbstractComponentLoader):
-            self._wrapped_command.load_into_component(component)
-
-
-class MessageCommandGroup(MessageCommand[abc.CommandCallbackSigT], abc.MessageCommandGroup[abc.CommandCallbackSigT]):
-    """Standard implementation of a message command group."""
-
-    __slots__ = ("_commands", "_is_strict", "_names_to_commands")
-
-    def __init__(
-        self,
-        callback: abc.CommandCallbackSigT,
-        name: str,
-        /,
-        *names: str,
-        strict: bool = False,
-        _wrapped_command: typing.Optional[abc.ExecutableCommand[typing.Any]] = None,
-    ) -> None:
-        """Initialise a message command group.
-
-        Parameters
-        ----------
-        name : str
-            The command name.
-
-        Other Parameters
-        ----------------
-        *names : str
-            Variable positional arguments of other names for the command.
-        strict : bool
-            Whether this command group should only allow commands without spaces in their names.
-
-            This allows for a more optimised command search pattern to be used and
-            enforces that command names are unique to a single command within the group.
-        """
-        super().__init__(callback, name, *names, _wrapped_command=_wrapped_command)
-        self._commands: list[abc.MessageCommand[typing.Any]] = []
-        self._is_strict = strict
-        self._names_to_commands: dict[str, abc.MessageCommand[typing.Any]] = {}
-
-    def __repr__(self) -> str:
-        return f"CommandGroup <{len(self._commands)}: {self._names}>"
-
-    @property
-    def commands(self) -> collections.Collection[abc.MessageCommand[typing.Any]]:
-        # <<inherited docstring from tanjun.abc.MessageCommandGroup>>.
-        return self._commands.copy()
-
-    @property
-    def is_strict(self) -> bool:
-        return self._is_strict
-
-    def copy(
-        self: _MessageCommandGroupT,
-        *,
-        parent: typing.Optional[abc.MessageCommandGroup[typing.Any]] = None,
-        _new: bool = True,
-    ) -> _MessageCommandGroupT:
-        # <<inherited docstring from tanjun.abc.MessageCommand>>.
-        if not _new:
-            commands = {command: command.copy(parent=self) for command in self._commands}
-            self._commands = list(commands.values())
-            self._names_to_commands = {name: commands[command] for name, command in self._names_to_commands.items()}
-            return super().copy(parent=parent, _new=_new)
-
-        return super().copy(parent=parent, _new=_new)
-
-    def add_command(self: _MessageCommandGroupT, command: abc.MessageCommand[typing.Any], /) -> _MessageCommandGroupT:
-        """Add a command to this group.
-
-        Parameters
-        ----------
-        command : MessageCommand
-            The command to add.
-
-        Returns
-        -------
-        Self
-            The group instance to enable chained calls.
-
-        Raises
-        ------
-        ValueError
-            If one of the command's names is already registered in a strict
-            command group.
-        """
-        if command in self._commands:
-            return self
-
-        if self._is_strict:
-            if any(" " in name for name in command.names):
-                raise ValueError("Sub-command names may not contain spaces in a strict message command group")
-
-            if name_conflicts := self._names_to_commands.keys() & command.names:
-                raise ValueError(
-                    "Sub-command names must be unique in a strict message command group. "
-                    "The following conflicts were found " + ", ".join(name_conflicts)
-                )
-
-            self._names_to_commands.update((name, command) for name in command.names)
-
-        command.set_parent(self)
-        self._commands.append(command)
-        return self
-
-    def remove_command(
-        self: _MessageCommandGroupT, command: abc.MessageCommand[typing.Any], /
-    ) -> _MessageCommandGroupT:
-        # <<inherited docstring from tanjun.abc.MessageCommandGroup>>.
-        self._commands.remove(command)
-        if self._is_strict:
-            for name in command.names:
-                if self._names_to_commands.get(name) == command:
-                    del self._names_to_commands[name]
-
-        command.set_parent(None)
-        return self
-
-    def with_command(self, command: AnyMessageCommandT, /) -> AnyMessageCommandT:
-        self.add_command(command)
-        return command
-
-    def bind_client(self: _MessageCommandGroupT, client: abc.Client, /) -> _MessageCommandGroupT:
-        # <<inherited docstring from tanjun.abc.ExecutableCommand>>.
-        super().bind_client(client)
-        for command in self._commands:
-            command.bind_client(client)
-
-        return self
-
-    def bind_component(self: _MessageCommandGroupT, component: abc.Component, /) -> _MessageCommandGroupT:
-        # <<inherited docstring from tanjun.abc.ExecutableCommand>>.
-        super().bind_component(component)
-        for command in self._commands:
-            command.bind_component(component)
-
-        return self
-
-    def find_command(self, content: str, /) -> collections.Iterable[tuple[str, abc.MessageCommand[typing.Any]]]:
-        if self._is_strict:
-            name = content.split(" ")[0]
-            if command := self._names_to_commands.get(name):
-                yield name, command
-            return
-
-        for command in self._commands:
-            if (name_ := utilities.match_prefix_names(content, command.names)) is not None:
-                yield name_, command
-
-    async def execute(
-        self,
-        ctx: abc.MessageContext,
-        /,
-        *,
-        hooks: typing.Optional[collections.MutableSet[abc.MessageHooks]] = None,
-    ) -> None:
-        # <<inherited docstring from tanjun.abc.MessageCommand>>.
-        if ctx.message.content is None:
-            raise ValueError("Cannot execute a command with a content-less message")
-
-        if self._hooks:
-            if hooks is None:
-                hooks = set()
-
-            hooks.add(self._hooks)
-
-        for name, command in self.find_command(ctx.content):
-            if await command.check_context(ctx):
-                content = ctx.content[len(name) :]
-                lstripped_content = content.lstrip()
-                space_len = len(content) - len(lstripped_content)
-                ctx.set_triggering_name(ctx.triggering_name + (" " * space_len) + name)
-                ctx.set_content(lstripped_content)
-                await command.execute(ctx, hooks=hooks)
-                return
-
-        await super().execute(ctx, hooks=hooks)
