+++ tests/commands/test_menu.py
@@ -29,11 +29,8 @@
 # CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-"""Standard implementations of Tanjun's context classes."""
-from __future__ import annotations
 
-__all__: list[str] = ["MessageContext", "SlashContext", "SlashOption"]
-
-from .message import MessageContext
-from .slash import SlashContext
-from .slash import SlashOption
+# pyright: reportIncompatibleMethodOverride=none
+# pyright: reportUnknownMemberType=none
+# pyright: reportPrivateUsage=none
+# This leads to too many false-positives around mocks.
