+++ tests/commands/test_slash.py
@@ -62,121 +62,6 @@ def stub_class(cls: type[_T], /, **namespace: typing.Any) -> type[_T]:
     return typing.cast(type[_T], new_cls)
 
 
-class TestPartialCommand:
-    @pytest.fixture()
-    def command(self) -> tanjun.commands.PartialCommand[typing.Any]:
-        fields: dict[str, typing.Any] = {}
-        for name in tanjun.commands.PartialCommand.__abstractmethods__:
-            fields[name] = mock.MagicMock()
-
-        return types.new_class(
-            "PartialCommand", (tanjun.commands.PartialCommand[typing.Any],), exec_body=lambda body: body.update(fields)
-        )()
-
-    def test_metadata_property(self, command: tanjun.commands.PartialCommand[typing.Any]):
-        assert command.metadata is command._metadata
-
-    def test_needs_injector_when_any_true(self, command: tanjun.commands.PartialCommand[typing.Any]):
-        command._checks = [mock.Mock(needs_injector=False), mock.Mock(needs_injector=True)]
-        assert command.needs_injector is True
-
-    def test_needs_injector_when_all_true(self, command: tanjun.commands.PartialCommand[typing.Any]):
-        command._checks = [mock.Mock(needs_injector=True), mock.Mock(needs_injector=True)]
-        assert command.needs_injector is True
-
-    def test_needs_injector_when_all_false(self, command: tanjun.commands.PartialCommand[typing.Any]):
-        command._checks = [mock.Mock(needs_injector=False), mock.Mock(needs_injector=False)]
-        assert command.needs_injector is False
-
-    def test_needs_injector_when_empty(self, command: tanjun.commands.PartialCommand[typing.Any]):
-        assert command.needs_injector is False
-
-    def test_copy(self, command: tanjun.commands.PartialCommand[typing.Any]):
-        mock_check = mock.Mock()
-        command._checks = [mock_check]
-        command._hooks = mock.Mock()
-        mock_metadata = mock.Mock()
-        command._metadata = mock_metadata
-
-        new_command = command.copy()
-
-        assert new_command is not command
-        new_command._checks == [mock_check.copy.return_value]
-        assert new_command._hooks is command._hooks.copy.return_value
-        assert new_command._metadata is mock_metadata.copy.return_value
-
-    def test_set_hooks(self, command: tanjun.commands.PartialCommand[typing.Any]):
-        mock_hooks = mock.Mock()
-
-        assert command.set_hooks(mock_hooks) is command
-        assert command.hooks is mock_hooks
-
-    def test_set_metadata(self, command: tanjun.commands.PartialCommand[typing.Any]):
-        key = mock.Mock()
-        value = mock.Mock()
-
-        result = command.set_metadata(key, value)
-
-        assert result is command
-        assert command.metadata[key] is value
-
-    def test_add_check(self, command: tanjun.commands.PartialCommand[typing.Any]):
-        mock_check = mock.Mock()
-
-        assert command.add_check(mock_check) is command
-
-        assert len(command._checks) == 1
-        check = next(iter(command._checks))
-        assert isinstance(check, tanjun.checks.InjectableCheck)
-        assert check.callback is mock_check
-        assert command.checks == (mock_check,)
-
-    def test_add_check_when_already_present(self, command: tanjun.commands.PartialCommand[typing.Any]):
-        mock_check = mock.Mock()
-
-        assert command.add_check(mock_check).add_check(mock_check) is command
-
-        assert list(command.checks).count(mock_check) == 1
-
-    def test_remove_check(self, command: tanjun.commands.PartialCommand[typing.Any]):
-        def mock_check() -> bool:
-            raise NotImplementedError
-
-        command._checks = [tanjun.checks.InjectableCheck(mock_check)]
-
-        result = command.remove_check(mock_check)
-
-        assert result is command
-        assert command.checks == ()
-
-    def test_with_check(self, command: tanjun.commands.PartialCommand[typing.Any]):
-        mock_check = mock.Mock()
-        add_check = mock.Mock()
-        command = stub_class(tanjun.commands.PartialCommand, add_check=add_check)()
-
-        assert command.with_check(mock_check) is mock_check
-        add_check.assert_called_once_with(mock_check)
-
-    def test_with_check_when_already_present(self, command: tanjun.commands.PartialCommand[typing.Any]):
-        def mock_check() -> bool:
-            raise NotImplementedError
-
-        command.add_check(mock_check).with_check(mock_check)
-        assert command.with_check(mock_check) is mock_check
-
-        assert list(command.checks).count(mock_check) == 1
-
-    def test_bind_client(self, command: tanjun.commands.PartialCommand[typing.Any]):
-        command.bind_client(mock.Mock())
-
-    def test_bind_component(self, command: tanjun.commands.PartialCommand[typing.Any]):
-        mock_component = mock.Mock()
-
-        command.bind_component(mock_component)
-
-        assert command.component is mock_component
-
-
 def test_slash_command_group():
     command = tanjun.slash_command_group(
         "a_name", "very", default_permission=False, default_to_ephemeral=True, is_global=False
@@ -289,7 +174,7 @@ def test_with_str_slash_option_with_defaults():
     mock_command.add_str_option.assert_called_once_with(
         "a_name",
         "a_value",
-        default=tanjun.commands._UNDEFINED_DEFAULT,
+        default=tanjun.commands.slash.UNDEFINED_DEFAULT,
         choices=None,
         converters=(),
         pass_as_kwarg=True,
@@ -337,7 +222,7 @@ def test_with_int_slash_option_with_defaults():
         "con man",
         choices=None,
         converters=(),
-        default=tanjun.commands._UNDEFINED_DEFAULT,
+        default=tanjun.commands.slash.UNDEFINED_DEFAULT,
         min_value=None,
         max_value=None,
         pass_as_kwarg=True,
@@ -386,7 +271,7 @@ def test_with_float_slash_option_with_defaults():
         "hi",
         "bye",
         always_float=True,
-        default=tanjun.commands._UNDEFINED_DEFAULT,
+        default=tanjun.commands.slash.UNDEFINED_DEFAULT,
         choices=None,
         converters=(),
         min_value=None,
@@ -412,7 +297,7 @@ def test_with_bool_slash_option_with_defaults():
 
     assert result is mock_command.add_bool_option.return_value
     mock_command.add_bool_option.assert_called_once_with(
-        "bool", "bool me man", default=tanjun.commands._UNDEFINED_DEFAULT, pass_as_kwarg=True
+        "bool", "bool me man", default=tanjun.commands.slash.UNDEFINED_DEFAULT, pass_as_kwarg=True
     )
 
 
@@ -436,7 +321,7 @@ def test_with_user_slash_option_with_defaults():
 
     assert result is mock_command.add_user_option.return_value
     mock_command.add_user_option.assert_called_once_with(
-        "victim", "who're we getting next?", default=tanjun.commands._UNDEFINED_DEFAULT, pass_as_kwarg=True
+        "victim", "who're we getting next?", default=tanjun.commands.slash.UNDEFINED_DEFAULT, pass_as_kwarg=True
     )
 
 
@@ -455,7 +340,9 @@ def test_with_member_slash_option_with_defaults():
     result = tanjun.with_member_slash_option("no", "hihihi?")(mock_command)
 
     assert result is mock_command.add_member_option.return_value
-    mock_command.add_member_option.assert_called_once_with("no", "hihihi?", default=tanjun.commands._UNDEFINED_DEFAULT)
+    mock_command.add_member_option.assert_called_once_with(
+        "no", "hihihi?", default=tanjun.commands.slash.UNDEFINED_DEFAULT
+    )
 
 
 def test_with_channel_slash_option():
@@ -478,7 +365,7 @@ def test_with_channel_slash_option_with_defaults():
 
     assert result is mock_command.add_channel_option.return_value
     mock_command.add_channel_option.assert_called_once_with(
-        "channel", "channel?", types=None, default=tanjun.commands._UNDEFINED_DEFAULT, pass_as_kwarg=True
+        "channel", "channel?", types=None, default=tanjun.commands.slash.UNDEFINED_DEFAULT, pass_as_kwarg=True
     )
 
 
@@ -498,7 +385,7 @@ def test_with_role_slash_option_with_defaults():
 
     assert result is mock_command.add_role_option.return_value
     mock_command.add_role_option.assert_called_once_with(
-        "role", "role?", default=tanjun.commands._UNDEFINED_DEFAULT, pass_as_kwarg=True
+        "role", "role?", default=tanjun.commands.slash.UNDEFINED_DEFAULT, pass_as_kwarg=True
     )
 
 
@@ -518,14 +405,14 @@ def test_with_mentionable_slash_option_with_defaults():
 
     assert result is mock_command.add_mentionable_option.return_value
     mock_command.add_mentionable_option.assert_called_once_with(
-        "mentu", "mentu?", default=tanjun.commands._UNDEFINED_DEFAULT, pass_as_kwarg=True
+        "mentu", "mentu?", default=tanjun.commands.slash.UNDEFINED_DEFAULT, pass_as_kwarg=True
     )
 
 
 class Test_TrackedOption:
     def test_init(self):
         mock_converter = mock.Mock()
-        option = tanjun.commands._TrackedOption(
+        option = tanjun.commands.slash._TrackedOption(
             name="name",
             option_type=hikari.OptionType.FLOAT,
             always_float=False,
@@ -542,7 +429,7 @@ class Test_TrackedOption:
         assert option.default == "default"
 
     def test_needs_injector_property_when_all_false(self):
-        option = tanjun.commands._TrackedOption(
+        option = tanjun.commands.slash._TrackedOption(
             name="no",
             option_type=hikari.OptionType.INTEGER,
             converters=[
@@ -555,12 +442,12 @@ class Test_TrackedOption:
         assert option.needs_injector is False
 
     def test_needs_injector_property_when_no_converters(self):
-        option = tanjun.commands._TrackedOption(name="no", option_type=hikari.OptionType.FLOAT)
+        option = tanjun.commands.slash._TrackedOption(name="no", option_type=hikari.OptionType.FLOAT)
 
         assert option.needs_injector is False
 
     def test_needs_injector_property_when_true(self):
-        option = tanjun.commands._TrackedOption(
+        option = tanjun.commands.slash._TrackedOption(
             name="no",
             option_type=hikari.OptionType.FLOAT,
             converters=[
@@ -575,7 +462,7 @@ class Test_TrackedOption:
     @pytest.mark.asyncio()
     async def test_convert_when_no_converters(self):
         mock_value = mock.Mock()
-        option = tanjun.commands._TrackedOption(name="hi", option_type=hikari.OptionType.INTEGER)
+        option = tanjun.commands.slash._TrackedOption(name="hi", option_type=hikari.OptionType.INTEGER)
 
         assert await option.convert(mock.Mock(), mock_value) is mock_value
 
@@ -589,7 +476,7 @@ class Test_TrackedOption:
         mock_converter_2.resolve_with_command_context.side_effect = exc_2
         mock_context = mock.Mock(base_context.BaseContext)
         mock_value = mock.Mock()
-        option = tanjun.commands._TrackedOption(
+        option = tanjun.commands.slash._TrackedOption(
             name="no", option_type=hikari.OptionType.FLOAT, converters=[mock_converter_1, mock_converter_2]
         )
 
@@ -610,7 +497,7 @@ class Test_TrackedOption:
         mock_converter_3 = mock.AsyncMock()
         mock_context = mock.Mock(base_context.BaseContext)
         mock_value = mock.Mock()
-        option = tanjun.commands._TrackedOption(
+        option = tanjun.commands.slash._TrackedOption(
             name="no",
             option_type=hikari.OptionType.FLOAT,
             converters=[mock_converter_1, mock_converter_2, mock_converter_3],
@@ -766,7 +653,7 @@ class TestSlashCommandGroup:
         result = command_group.build()
 
         assert result == (
-            tanjun.commands._SlashCommandBuilder("yee", "nsoosos", False)
+            tanjun.commands.slash._SlashCommandBuilder("yee", "nsoosos", False)
             .set_default_permission(True)
             .add_option(
                 hikari.CommandOption(
@@ -1052,7 +939,7 @@ class TestSlashCommand:
         tracked = command._tracked_options[option.name]
         assert tracked.name == option.name
         assert tracked.type is hikari.OptionType.STRING
-        assert tracked.default is tanjun.commands._UNDEFINED_DEFAULT
+        assert tracked.default is tanjun.commands.slash.UNDEFINED_DEFAULT
         assert tracked.converters == []
         assert tracked.is_always_float is False
         assert tracked.is_only_member is False
@@ -1086,7 +973,7 @@ class TestSlashCommand:
         tracked = command._tracked_options[option.name]
         assert tracked.name == option.name
         assert tracked.type is hikari.OptionType.STRING
-        assert tracked.default is tanjun.commands._UNDEFINED_DEFAULT
+        assert tracked.default is tanjun.commands.slash.UNDEFINED_DEFAULT
         assert tracked.converters == []
         assert tracked.is_always_float is False
         assert tracked.is_only_member is False
@@ -1108,7 +995,7 @@ class TestSlashCommand:
         tracked = command._tracked_options[option.name]
         assert tracked.name == option.name
         assert tracked.type is hikari.OptionType.STRING
-        assert tracked.default is tanjun.commands._UNDEFINED_DEFAULT
+        assert tracked.default is tanjun.commands.slash.UNDEFINED_DEFAULT
         assert tracked.converters == []
         assert tracked.is_always_float is False
         assert tracked.is_only_member is False
@@ -1237,7 +1124,7 @@ class TestSlashCommand:
         tracked = command._tracked_options[option.name]
         assert tracked.name == option.name
         assert tracked.type is hikari.OptionType.INTEGER
-        assert tracked.default is tanjun.commands._UNDEFINED_DEFAULT
+        assert tracked.default is tanjun.commands.slash.UNDEFINED_DEFAULT
         assert tracked.converters == []
         assert tracked.is_always_float is False
         assert tracked.is_only_member is False
@@ -1259,7 +1146,7 @@ class TestSlashCommand:
         tracked = command._tracked_options[option.name]
         assert tracked.name == option.name
         assert tracked.type is hikari.OptionType.INTEGER
-        assert tracked.default is tanjun.commands._UNDEFINED_DEFAULT
+        assert tracked.default is tanjun.commands.slash.UNDEFINED_DEFAULT
         assert tracked.converters == []
         assert tracked.is_always_float is False
         assert tracked.is_only_member is False
@@ -1412,7 +1299,7 @@ class TestSlashCommand:
         tracked = command._tracked_options[option.name]
         assert tracked.name == option.name
         assert tracked.type is hikari.OptionType.FLOAT
-        assert tracked.default is tanjun.commands._UNDEFINED_DEFAULT
+        assert tracked.default is tanjun.commands.slash.UNDEFINED_DEFAULT
         assert tracked.converters == []
         assert tracked.is_always_float is True
         assert tracked.is_only_member is False
@@ -1434,7 +1321,7 @@ class TestSlashCommand:
         tracked = command._tracked_options[option.name]
         assert tracked.name == option.name
         assert tracked.type is hikari.OptionType.FLOAT
-        assert tracked.default is tanjun.commands._UNDEFINED_DEFAULT
+        assert tracked.default is tanjun.commands.slash.UNDEFINED_DEFAULT
         assert tracked.converters == []
         assert tracked.is_always_float is True
         assert tracked.is_only_member is False
@@ -1529,7 +1416,7 @@ class TestSlashCommand:
         tracked = command._tracked_options[option.name]
         assert tracked.name == option.name
         assert tracked.type is hikari.OptionType.BOOLEAN
-        assert tracked.default is tanjun.commands._UNDEFINED_DEFAULT
+        assert tracked.default is tanjun.commands.slash.UNDEFINED_DEFAULT
         assert tracked.converters == []
         assert tracked.is_always_float is False
         assert tracked.is_only_member is False
@@ -1621,7 +1508,7 @@ class TestSlashCommand:
         tracked = command._tracked_options[option.name]
         assert tracked.name == option.name
         assert tracked.type is hikari.OptionType.USER
-        assert tracked.default is tanjun.commands._UNDEFINED_DEFAULT
+        assert tracked.default is tanjun.commands.slash.UNDEFINED_DEFAULT
         assert tracked.converters == []
         assert tracked.is_always_float is False
         assert tracked.is_only_member is False
@@ -1713,7 +1600,7 @@ class TestSlashCommand:
         tracked = command._tracked_options[option.name]
         assert tracked.name == option.name
         assert tracked.type is hikari.OptionType.USER
-        assert tracked.default is tanjun.commands._UNDEFINED_DEFAULT
+        assert tracked.default is tanjun.commands.slash.UNDEFINED_DEFAULT
         assert tracked.converters == []
         assert tracked.is_always_float is False
         assert tracked.is_only_member is True
@@ -1794,7 +1681,7 @@ class TestSlashCommand:
         tracked = command._tracked_options[option.name]
         assert tracked.name == option.name
         assert tracked.type is hikari.OptionType.CHANNEL
-        assert tracked.default is tanjun.commands._UNDEFINED_DEFAULT
+        assert tracked.default is tanjun.commands.slash.UNDEFINED_DEFAULT
         assert tracked.converters == []
         assert tracked.is_always_float is False
         assert tracked.is_only_member is False
@@ -1914,7 +1801,7 @@ class TestSlashCommand:
         tracked = command._tracked_options[option.name]
         assert tracked.name == option.name
         assert tracked.type is hikari.OptionType.ROLE
-        assert tracked.default is tanjun.commands._UNDEFINED_DEFAULT
+        assert tracked.default is tanjun.commands.slash.UNDEFINED_DEFAULT
         assert tracked.converters == []
         assert tracked.is_always_float is False
         assert tracked.is_only_member is False
@@ -2003,7 +1890,7 @@ class TestSlashCommand:
         tracked = command._tracked_options[option.name]
         assert tracked.name == option.name
         assert tracked.type is hikari.OptionType.MENTIONABLE
-        assert tracked.default is tanjun.commands._UNDEFINED_DEFAULT
+        assert tracked.default is tanjun.commands.slash.UNDEFINED_DEFAULT
         assert tracked.converters == []
         assert tracked.is_always_float is False
         assert tracked.is_only_member is False
@@ -2069,574 +1956,3 @@ class TestSlashCommand:
     @pytest.mark.skip(reason="TODO")
     def test_copy(self):
         ...
-
-
-def test_as_message_command():
-    mock_callback = mock.Mock()
-    command = tanjun.as_message_command("a", "b")(mock_callback)
-
-    assert command.names == ["a", "b"]
-    assert command.callback is mock_callback
-    assert command._wrapped_command is None
-
-
-@pytest.mark.parametrize(
-    "other_command",
-    [
-        tanjun.SlashCommand(mock.Mock(), "e", "a"),
-        tanjun.MessageCommand(mock.Mock(), "b"),
-        tanjun.MessageCommandGroup(mock.Mock(), "b"),
-    ],
-)
-def test_as_message_command_when_wrapping_command(
-    other_command: typing.Union[
-        tanjun.SlashCommand[typing.Any], tanjun.MessageCommand[typing.Any], tanjun.MessageCommandGroup[typing.Any]
-    ]
-):
-    command = tanjun.as_message_command("a", "b")(other_command)
-
-    assert command._wrapped_command is other_command
-
-
-def test_as_message_command_group():
-    mock_callback = mock.Mock()
-    command = tanjun.as_message_command_group("c", "b", strict=True)(mock_callback)
-
-    assert command.names == ["c", "b"]
-    assert command.is_strict is True
-    assert command.callback is mock_callback
-    assert command._wrapped_command is None
-
-
-@pytest.mark.parametrize(
-    "other_command",
-    [
-        tanjun.SlashCommand(mock.Mock(), "e", "a"),
-        tanjun.MessageCommand(mock.Mock(), "b"),
-        tanjun.MessageCommandGroup(mock.Mock(), "b"),
-    ],
-)
-def test_as_message_command_group_when_wrapping_command(
-    other_command: typing.Union[
-        tanjun.SlashCommand[typing.Any], tanjun.MessageCommand[typing.Any], tanjun.MessageCommandGroup[typing.Any]
-    ]
-):
-    command = tanjun.as_message_command_group("c", "b", strict=True)(other_command)
-
-    assert command._wrapped_command is other_command
-
-
-class TestMessageCommand:
-    @pytest.mark.parametrize(
-        "inner_command", [tanjun.SlashCommand(mock.Mock(), "a", "b"), tanjun.MessageCommand(mock.Mock(), "a")]
-    )
-    def test___init___when_command_object(
-        self,
-        inner_command: typing.Union[
-            tanjun.SlashCommand[tanjun.abc.CommandCallbackSig], tanjun.MessageCommand[tanjun.abc.CommandCallbackSig]
-        ],
-    ):
-        assert tanjun.MessageCommand(inner_command, "woow").callback is inner_command.callback
-
-    @pytest.mark.skip(reason="TODO")
-    def test___repr__(self):
-        ...
-
-    @pytest.mark.asyncio()
-    async def test___call__(self):
-        mock_callback = mock.AsyncMock()
-        command = tanjun.SlashCommand(mock_callback, "yee", "nsoosos")
-
-        await command(65123, "okokok", a="odoosd", gf=435123)  # type: ignore
-
-        mock_callback.assert_awaited_once_with(65123, "okokok", a="odoosd", gf=435123)
-
-    def test_callback_property(self):
-        mock_callback = mock.Mock()
-
-        assert tanjun.MessageCommand(mock_callback, "yee", "nsoosos").callback is mock_callback
-
-    def test_names_property(self):
-        command = tanjun.MessageCommand(mock.Mock(), "aaaaa", "bbbbb", "ccccc")
-
-        assert command.names == ["aaaaa", "bbbbb", "ccccc"]
-
-    @pytest.mark.skip(reason="TODO")
-    def test_needs_injector_property(self):
-        ...
-
-    def test_parent_property(self):
-        mock_parent = mock.Mock()
-        command = tanjun.MessageCommand(mock.Mock(), "aaaaa", "bbbbb", "ccccc").set_parent(mock_parent)
-
-        assert command.parent is mock_parent
-
-    def test_parser_property(self):
-        mock_parser = mock.Mock()
-        command = tanjun.MessageCommand(mock.Mock(), "aaaaa", "bbbbb", "ccccc").set_parser(mock_parser)
-
-        assert command.parser is mock_parser
-
-    def test_bind_client(self):
-        mock_client = mock.Mock()
-        command = tanjun.MessageCommand(mock.Mock(), "aaaaa", "bbbbb", "ccccc")
-
-        with mock.patch.object(tanjun.commands.PartialCommand, "bind_client") as bind_client:
-            command.bind_client(mock_client)
-
-            bind_client.assert_called_once_with(mock_client)
-
-    def test_bind_client_when_has_parser(self):
-        mock_client = mock.Mock()
-        mock_parser = mock.Mock()
-        command = tanjun.MessageCommand(mock.Mock(), "aaaaa", "bbbbb", "ccccc").set_parser(mock_parser)
-
-        with mock.patch.object(tanjun.commands.PartialCommand, "bind_client") as bind_client:
-            command.bind_client(mock_client)
-
-            bind_client.assert_called_once_with(mock_client)
-
-        mock_parser.bind_client.assert_called_once_with(mock_client)
-
-    def test_bind_component(self):
-        mock_component = mock.Mock()
-        command = tanjun.MessageCommand(mock.Mock(), "aaaaa", "bbbbb", "ccccc")
-
-        with mock.patch.object(tanjun.commands.PartialCommand, "bind_component") as bind_component:
-            command.bind_component(mock_component)
-
-            bind_component.assert_called_once_with(mock_component)
-
-    def test_bind_component_when_has_parser(self):
-        mock_component = mock.Mock()
-        mock_parser = mock.Mock()
-        command = tanjun.MessageCommand(mock.Mock(), "aaaaa", "bbbbb", "ccccc").set_parser(mock_parser)
-
-        with mock.patch.object(tanjun.commands.PartialCommand, "bind_component") as bind_component:
-            command.bind_component(mock_component)
-
-            bind_component.assert_called_once_with(mock_component)
-
-        mock_parser.bind_component.assert_called_once_with(mock_component)
-
-    @pytest.mark.skip(reason="TODO")
-    def test_copy(self):
-        ...
-
-    @pytest.mark.asyncio()
-    async def test_check_context(self):
-        mock_callback = mock.Mock()
-        mock_other_callback = mock.Mock()
-        mock_context = mock.Mock()
-        mock_checks = [mock.Mock(), mock.Mock()]
-
-        with mock.patch.object(tanjun.checks, "InjectableCheck", side_effect=mock_checks.copy()) as injectable_check:
-            command = (
-                tanjun.MessageCommand(mock.Mock(), "yee", "nsoosos")
-                .add_check(mock_callback)
-                .add_check(mock_other_callback)
-            )
-
-            injectable_check.call_args_list == [mock.call(mock_callback), mock.call(mock_other_callback)]
-
-        with mock.patch.object(tanjun.utilities, "gather_checks", new=mock.AsyncMock()) as gather_checks:
-            result = await command.check_context(mock_context)
-
-            gather_checks.assert_awaited_once_with(mock_context, mock_checks)
-
-        assert result is gather_checks.return_value
-        mock_context.set_command.assert_has_calls([mock.call(command), mock.call(None)])
-
-    @pytest.mark.skip(reason="TODO")
-    @pytest.mark.asyncio()
-    async def test_execute(self):
-        ...
-
-    def test_load_into_component(self):
-        mock_component = mock.Mock()
-        command = tanjun.MessageCommand(mock.Mock(), "yee", "nsoosos")
-
-        command.load_into_component(mock_component)
-
-        mock_component.add_message_command.assert_called_once_with(command)
-
-    def test_load_into_component_when_wrapped_command_set(self):
-        mock_component = mock.Mock()
-        mock_other_command = mock.Mock()
-        command = tanjun.MessageCommand(mock.Mock(), "yee", "nsoosos")
-        command._wrapped_command = mock_other_command
-
-        command.load_into_component(mock_component)
-
-        mock_component.add_message_command.assert_called_once_with(command)
-        mock_other_command.load_into_component.assert_not_called()
-
-    def test_load_into_component_when_wrapped_command_is_loadable(self):
-        mock_component = mock.Mock()
-        mock_other_command = mock.Mock(tanjun.components.AbstractComponentLoader)
-        command = tanjun.MessageCommand(mock.Mock(), "yee", "nsoosos")
-        command._wrapped_command = mock_other_command
-
-        command.load_into_component(mock_component)
-
-        mock_component.add_message_command.assert_called_once_with(command)
-        mock_other_command.load_into_component.assert_called_once_with(mock_component)
-
-
-class TestMessageCommandGroup:
-    @pytest.mark.skip(reason="TODO")
-    def test___repr__(self):
-        ...
-
-    @pytest.mark.skip(reason="TODO")
-    def test_commands_property(self):
-        ...
-
-    @pytest.mark.skip(reason="TODO")
-    def test_is_strict_property(self):
-        ...
-
-    @pytest.mark.skip(reason="TODO")
-    def test_copy(self):
-        ...
-
-    def test_add_command(self):
-        mock_command = mock.Mock()
-        command_group = tanjun.MessageCommandGroup(mock.Mock(), "yee", "nsoosos")
-
-        result = command_group.add_command(mock_command)
-
-        assert result is command_group
-        mock_command.set_parent.assert_called_once_with(command_group)
-        assert mock_command in command_group.commands
-
-    def test_add_command_when_already_present(self):
-        mock_command = mock.Mock()
-        command_group = tanjun.MessageCommandGroup(mock.Mock(), "yee", "nsoosos")
-
-        result = command_group.add_command(mock_command).add_command(mock_command)
-
-        assert result is command_group
-        assert list(command_group.commands).count(mock_command) == 1
-
-    def test_add_command_when_strict(self):
-        mock_command = mock.Mock(names={"a", "b"})
-        command_group = tanjun.MessageCommandGroup(mock.Mock(), "yee", "nsoosos", strict=True)
-
-        result = command_group.add_command(mock_command)
-
-        assert result is command_group
-        mock_command.set_parent.assert_called_once_with(command_group)
-        assert mock_command in command_group.commands
-
-    def test_add_command_when_strict_and_space_in_any_name(self):
-        command_group = tanjun.MessageCommandGroup(mock.Mock(), "yee", "nsoosos", strict=True)
-
-        with pytest.raises(
-            ValueError, match="Sub-command names may not contain spaces in a strict message command group"
-        ):
-            command_group.add_command(mock.Mock(names={"a space", "b"}))
-
-    def test_add_command_when_strict_and_conflicts_found(self):
-        command_group = (
-            tanjun.MessageCommandGroup(mock.Mock(), "yee", "nsoosos", strict=True)
-            .add_command(mock.Mock(names={"aaa", "b"}))
-            .add_command(mock.Mock(names={"fsa", "dsaasd"}))
-        )
-
-        with pytest.raises(
-            ValueError,
-            match="Sub-command names must be unique in a strict message command group. "
-            "The following conflicts were found (?:aaa, dsaasd)|(?:dsaasd, aaa)",
-        ):
-            command_group.add_command(mock.Mock(names={"aaa", "dsaasd"}))
-
-    def test_remove_command(self):
-        mock_command = mock.Mock()
-        command_group = tanjun.MessageCommandGroup(mock.Mock(), "a", "b").add_command(mock_command)
-
-        result = command_group.remove_command(mock_command)
-
-        assert result is command_group
-        assert mock_command not in command_group.commands
-
-    def test_remove_command_when_strict(self):
-        mock_command = mock.Mock(names={"abba", "bba", "dadaba"})
-        mock_other_command = mock.Mock(names={"dada"})
-        command_group = (
-            tanjun.MessageCommandGroup(mock.Mock(), "a", "b", strict=True)
-            .add_command(mock_command)
-            .add_command(mock_other_command)
-        )
-
-        result = command_group.remove_command(mock_command)
-
-        assert result is command_group
-        assert mock_command not in command_group.commands
-        assert command_group._names_to_commands == {"dada": mock_other_command}
-
-    def test_with_command(self):
-        add_command = mock.Mock()
-        command = stub_class(tanjun.MessageCommandGroup[typing.Any], add_command=add_command)(mock.Mock(), "a", "b")
-        mock_command = mock.Mock()
-
-        result = command.with_command(mock_command)
-
-        assert result is mock_command
-        add_command.assert_called_once_with(mock_command)
-
-    def test_bind_client(self):
-        mock_command_1 = mock.Mock()
-        mock_command_2 = mock.Mock()
-        mock_command_3 = mock.Mock()
-        command = (
-            tanjun.MessageCommandGroup(mock.Mock(), "a", "b")
-            .add_command(mock_command_1)
-            .add_command(mock_command_2)
-            .add_command(mock_command_3)
-        )
-        mock_client = mock.Mock()
-
-        with mock.patch.object(tanjun.MessageCommand, "bind_client") as bind_client:
-            command.bind_client(mock_client)
-
-            bind_client.assert_called_once_with(mock_client)
-
-        mock_command_1.bind_client.assert_called_once_with(mock_client)
-        mock_command_2.bind_client.assert_called_once_with(mock_client)
-        mock_command_3.bind_client.assert_called_once_with(mock_client)
-
-    def test_bind_component(self):
-        mock_command_1 = mock.Mock()
-        mock_command_2 = mock.Mock()
-        mock_command_3 = mock.Mock()
-        command = (
-            tanjun.MessageCommandGroup(mock.Mock(), "a", "b")
-            .add_command(mock_command_1)
-            .add_command(mock_command_2)
-            .add_command(mock_command_3)
-        )
-        mock_component = mock.Mock()
-
-        with mock.patch.object(tanjun.MessageCommand, "bind_component") as bind_component:
-            command.bind_component(mock_component)
-
-            bind_component.assert_called_once_with(mock_component)
-
-        mock_command_1.bind_component.assert_called_once_with(mock_component)
-        mock_command_2.bind_component.assert_called_once_with(mock_component)
-        mock_command_3.bind_component.assert_called_once_with(mock_component)
-
-    def test_find_command(self):
-        mock_command_1 = mock.Mock(names={"i am", "sexy", "jk", "unless"})
-        mock_command_2 = mock.Mock(names={"i", "owo uwu", "no u"})
-        command_group = (
-            tanjun.MessageCommandGroup(mock.Mock(), "yee", "nsoosos")
-            .add_command(mock_command_1)
-            .add_command(mock.Mock(names={"ok boomer", "no u"}))
-            .add_command(mock_command_2)
-            .add_command(mock.Mock(names={"go home", "no u"}))
-        )
-
-        results = set(command_group.find_command("i am going home now"))
-
-        assert results == {("i", mock_command_2), ("i am", mock_command_1)}
-
-    def test_find_command_when_strict(self):
-        mock_command_1 = mock.Mock(names={"ok", "no"})
-        command_group = (
-            tanjun.MessageCommandGroup(mock.Mock(), "yee", "nsoosos", strict=True)
-            .add_command(mock_command_1)
-            .add_command(mock.Mock(names={"boomer"}))
-            .add_command(mock.Mock(names={"go"}))
-        )
-
-        results = list(command_group.find_command("ok i am going home now"))
-
-        assert results == [("ok", mock_command_1)]
-
-    def test_find_command_when_strict_and_unknown_name(self):
-        command_group = (
-            tanjun.MessageCommandGroup(mock.Mock(), "yee", "nsoosos", strict=True)
-            .add_command(mock.Mock(names={"ok", "no"}))
-            .add_command(mock.Mock(names={"boomer"}))
-            .add_command(mock.Mock(names={"go"}))
-        )
-
-        results = list(command_group.find_command("nobob i am going home now"))
-
-        assert results == []
-
-    @pytest.mark.asyncio()
-    async def test_execute_no_message_content(self):
-        mock_context = mock.Mock()
-        mock_context.message.content = None
-        command_group = tanjun.MessageCommandGroup(mock.AsyncMock(), "hi", "nsoosos")
-
-        with pytest.raises(ValueError, match="Cannot execute a command with a content-less message"):
-            await command_group.execute(mock_context)
-
-    @pytest.mark.asyncio()
-    async def test_execute(self):
-        mock_command_1 = mock.AsyncMock()
-        mock_command_1.check_context.return_value = False
-        mock_command_2 = mock.AsyncMock()
-        mock_command_2.check_context.return_value = True
-        mock_command_3 = mock.AsyncMock()
-        mock_command_3.check_context.return_value = True
-        mock_hooks = mock.Mock()
-        mock_context = mock.Mock(content="baka desu-ga hi", triggering_name="go home")
-        mock_attached_hooks = mock.Mock()
-        command = stub_class(
-            tanjun.MessageCommandGroup[typing.Any],
-            find_command=mock.Mock(
-                return_value=iter(
-                    [("onii-chan>////<", mock_command_1), ("baka", mock_command_2), ("nope", mock_command_3)]
-                )
-            ),
-        )(mock.AsyncMock(), "a", "b").set_hooks(mock_attached_hooks)
-
-        await command.execute(mock_context, hooks={typing.cast(tanjun.abc.MessageHooks, mock_hooks)})
-
-        mock_context.set_content.assert_called_once_with("desu-ga hi")
-        mock_context.set_triggering_name.assert_called_once_with("go home baka")
-        mock_command_1.check_context.assert_awaited_once_with(mock_context)
-        mock_command_2.check_context.assert_awaited_once_with(mock_context)
-        mock_command_3.check_context.assert_not_called()
-        mock_command_1.execute.assert_not_called()
-        mock_command_2.execute.assert_called_once_with(mock_context, hooks={mock_hooks, mock_attached_hooks})
-        mock_command_3.execute.assert_not_called()
-
-    @pytest.mark.asyncio()
-    async def test_execute_no_pass_through_hooks(self):
-        mock_command_1 = mock.AsyncMock()
-        mock_command_1.check_context.return_value = False
-        mock_command_2 = mock.AsyncMock()
-        mock_command_2.check_context.return_value = True
-        mock_command_3 = mock.AsyncMock()
-        mock_command_3.check_context.return_value = True
-        mock_context = mock.Mock(content="baka desu-ga hi", triggering_name="go home")
-        mock_attached_hooks = mock.Mock()
-        command = stub_class(
-            tanjun.MessageCommandGroup[typing.Any],
-            find_command=mock.Mock(
-                return_value=iter(
-                    [("onii-chan>////<", mock_command_1), ("baka", mock_command_2), ("nope", mock_command_3)]
-                )
-            ),
-        )(mock.AsyncMock(), "a", "b").set_hooks(mock_attached_hooks)
-
-        await command.execute(mock_context)
-
-        mock_context.set_content.assert_called_once_with("desu-ga hi")
-        mock_context.set_triggering_name.assert_called_once_with("go home baka")
-        mock_command_1.check_context.assert_awaited_once_with(mock_context)
-        mock_command_2.check_context.assert_awaited_once_with(mock_context)
-        mock_command_3.check_context.assert_not_called()
-        mock_command_1.execute.assert_not_called()
-        mock_command_2.execute.assert_called_once_with(mock_context, hooks={mock_attached_hooks})
-        mock_command_3.execute.assert_not_called()
-
-    @pytest.mark.asyncio()
-    async def test_execute_no_hooks(self):
-        mock_command_1 = mock.AsyncMock()
-        mock_command_1.check_context.return_value = False
-        mock_command_2 = mock.AsyncMock()
-        mock_command_2.check_context.return_value = True
-        mock_command_3 = mock.AsyncMock()
-        mock_command_3.check_context.return_value = True
-        mock_context = mock.Mock(content="baka desu-ga hi", triggering_name="go home")
-        command = stub_class(
-            tanjun.MessageCommandGroup[typing.Any],
-            find_command=mock.Mock(
-                return_value=iter(
-                    [("onii-chan>////<", mock_command_1), ("baka", mock_command_2), ("nope", mock_command_3)]
-                )
-            ),
-        )(mock.AsyncMock(), "a", "b")
-
-        await command.execute(mock_context)
-
-        mock_context.set_content.assert_called_once_with("desu-ga hi")
-        mock_context.set_triggering_name.assert_called_once_with("go home baka")
-        mock_command_1.check_context.assert_awaited_once_with(mock_context)
-        mock_command_2.check_context.assert_awaited_once_with(mock_context)
-        mock_command_3.check_context.assert_not_called()
-        mock_command_1.execute.assert_not_called()
-        mock_command_2.execute.assert_called_once_with(mock_context, hooks=None)
-        mock_command_3.execute.assert_not_called()
-
-    @pytest.mark.asyncio()
-    async def test_execute_falls_back_to_own_callback(self):
-        mock_command_1 = mock.AsyncMock()
-        mock_command_1.check_context.return_value = False
-        mock_command_2 = mock.AsyncMock()
-        mock_command_2.check_context.return_value = False
-        mock_hooks = mock.Mock()
-        mock_context = mock.Mock(content="baka desu-ga hi", triggering_name="go home")
-        mock_attached_hooks = mock.Mock()
-        command = stub_class(
-            tanjun.MessageCommandGroup[typing.Any],
-            find_command=mock.Mock(return_value=iter([("onii-chan>////<", mock_command_1), ("baka", mock_command_2)])),
-        )(mock.AsyncMock(), "a", "b").set_hooks(mock_attached_hooks)
-
-        with mock.patch.object(tanjun.commands.MessageCommand, "execute", new=mock.AsyncMock()) as mock_execute:
-            await command.execute(mock_context, hooks={typing.cast(tanjun.abc.MessageHooks, mock_hooks)})
-
-            mock_execute.assert_called_once_with(mock_context, hooks={mock_hooks, mock_attached_hooks})
-
-        mock_context.set_content.assert_not_called()
-        mock_context.set_triggering_name.assert_not_called()
-        mock_command_1.check_context.assert_awaited_once_with(mock_context)
-        mock_command_2.check_context.assert_awaited_once_with(mock_context)
-        mock_command_1.execute.assert_not_called()
-        mock_command_2.execute.assert_not_called()
-
-    @pytest.mark.asyncio()
-    async def test_execute_falls_back_to_own_callback_no_pass_through_hooks(self):
-        mock_command_1 = mock.AsyncMock()
-        mock_command_1.check_context.return_value = False
-        mock_command_2 = mock.AsyncMock()
-        mock_command_2.check_context.return_value = False
-        mock_context = mock.Mock(content="baka desu-ga hi", triggering_name="go home")
-        mock_attached_hooks = mock.Mock()
-        command = stub_class(
-            tanjun.MessageCommandGroup[typing.Any],
-            find_command=mock.Mock(return_value=iter([("onii-chan>////<", mock_command_1), ("baka", mock_command_2)])),
-        )(mock.AsyncMock(), "a", "b").set_hooks(mock_attached_hooks)
-
-        with mock.patch.object(tanjun.commands.MessageCommand, "execute", new=mock.AsyncMock()) as mock_execute:
-            await command.execute(mock_context)
-
-            mock_execute.assert_called_once_with(mock_context, hooks={mock_attached_hooks})
-
-        mock_context.set_content.assert_not_called()
-        mock_context.set_triggering_name.assert_not_called()
-        mock_command_1.check_context.assert_awaited_once_with(mock_context)
-        mock_command_2.check_context.assert_awaited_once_with(mock_context)
-        mock_command_1.execute.assert_not_called()
-        mock_command_2.execute.assert_not_called()
-
-    @pytest.mark.asyncio()
-    async def test_execute_falls_back_to_own_callback_no_hooks(self):
-        mock_command_1 = mock.AsyncMock()
-        mock_command_1.check_context.return_value = False
-        mock_command_2 = mock.AsyncMock()
-        mock_command_2.check_context.return_value = False
-        mock_context = mock.Mock(content="baka desu-ga hi", triggering_name="go home")
-        command = stub_class(
-            tanjun.MessageCommandGroup[typing.Any],
-            find_command=mock.Mock(return_value=iter([("onii-chan>////<", mock_command_1), ("baka", mock_command_2)])),
-        )(mock.AsyncMock(), "a", "b")
-
-        with mock.patch.object(tanjun.commands.MessageCommand, "execute", new=mock.AsyncMock()) as mock_execute:
-            await command.execute(mock_context)
-
-            mock_execute.assert_called_once_with(mock_context, hooks=None)
-
-        mock_context.set_content.assert_not_called()
-        mock_context.set_triggering_name.assert_not_called()
-        mock_command_1.check_context.assert_awaited_once_with(mock_context)
-        mock_command_2.check_context.assert_awaited_once_with(mock_context)
-        mock_command_1.execute.assert_not_called()
-        mock_command_2.execute.assert_not_called()
